+++ app/javascript/components/helpers.js
@@ -1,17 +1,19 @@
+import axios from "axios";
+
 export const fillCart = (cart, products) => {
   const filledCart = [];
-  Object.keys(cart).forEach(key => {
-    const myProduct = products.find(product => product.id == key);
+  Object.keys(cart).forEach((key) => {
+    const myProduct = products.find((product) => product.id == key);
     if (myProduct) {
       filledCart.push({
         id: key,
         name: myProduct.name,
         imageUrl: myProduct.image_tiny_url,
         price_cents: myProduct.price_cents,
-        quantity: cart[key]
+        quantity: cart[key],
       });
     } else {
-      console.log(`Unable to find product with ID ${key}`)
+      console.log(`Unable to find product with ID ${key}`);
     }
   });
   return filledCart;
@@ -28,4 +30,31 @@ export const formattedPrice = function (price_cents) {
   } else {
     return `$0.0${cents}`;
   }
+};
+
+export const submitOrder = async (cart) => {
+  try {
+    const simpleCart = cart.map((item) => {
+      return {
+        id: item.id,
+        price_cents: item.price_cents,
+        quantity: item.quantity,
+      };
+    });
+    console.log(simpleCart);
+    const sendOrder = await axios.post("/cart/fill", simpleCart);
+    if (sendOrder.data === "OK") {
+      window.location.href = `/cart`;
+    }
+  } catch (err) {
+    console.log("Errror: ", err.message);
+  }
+};
+
+export const subtotal = (cart) => {
+  return formattedPrice(
+    cart.reduce((acc, item) => {
+      return acc + item.price_cents * item.quantity;
+    }, 0)
+  );
 };
